:sparkles: Changed Seeding Data

+++ Backend/sequelize/seeders/4_A_User_Role.js
@@ -16,10 +16,25 @@ module.exports = {
          * }], {});
          */
         return queryInterface.bulkInsert('A_User_Role', [{
+            userUserId: constants.userIDAdmin,
+            roleRoleId: constants.roleIDAdmin,
             createdAt: new Date,
-            updatedAt: new Date,
-            userUserId: constants.userID,
-            roleRoleId: constants.roleID
+            updatedAt: new Date
+        },{
+            userUserId: constants.userIDTeacher,
+            roleRoleId: constants.roleIDTeacher,
+            createdAt: new Date,
+            updatedAt: new Date
+        },{
+            userUserId: constants.userIDStudent,
+            roleRoleId: constants.roleIDStudent,
+            createdAt: new Date,
+            updatedAt: new Date
+        },{
+            userUserId: constants.userIDUser,
+            roleRoleId: constants.roleIDUser,
+            createdAt: new Date,
+            updatedAt: new Date
         }]);
     },
 
@@ -30,9 +45,6 @@ module.exports = {
          * Example:
          * await queryInterface.bulkDelete('People', null, {});
          */
-        return queryInterface.bulkDelete('A_User_Role', {
-            userUserId: constants.userID,
-            roleRoleId: constants.roleID
-        })
+        return queryInterface.bulkDelete('A_User_Role', {user_id: {[Op.in]: [constants.userIDAdmin, constants.userIDTeacher, constants.userIDStudent, constants.userIDUser]}})
     }
 };
